Moved sliders
Moved x0 slider to the x-axis. Moved a slider to the y-axis. This can
change but I think it is the simplest solution (it would make it much
more complex to have the slider in a separate box). Arrows cannot be
added in the middle of the lines so I did not add any.